+++ lib/src/components/components.dart
@@ -32,4 +32,5 @@ export 'flowbite_input_label.dart';
 export 'flowbite_jumbotron.dart';
 export 'flowbite_outlined_button.dart';
 export 'flowbite_spinner.dart';
+export 'flowbite_toast.dart';
 export 'flowbite_toggle_input.dart';